+++ global_trotter_backend/app.py
@@ -8,6 +8,7 @@ from routes.comments import comments_bp
 from routes.destinations import destinations_bp
 from routes.recommendations import recommendations_bp
 from routes.itineraries import itineraries_bp
+from routes.uploads import uploads_bp
 
 
 def create_app():
@@ -22,8 +23,7 @@ def create_app():
     app.register_blueprint(destinations_bp)
     app.register_blueprint(recommendations_bp)
     app.register_blueprint(itineraries_bp)
-
-
+    app.register_blueprint(uploads_bp)
 
     return app